+++ Strategic/Creature Spreading.cpp
@@ -139,11 +139,11 @@ INT32 giDestroyedLairID = 0;
 //prebattle interface, autoresolve, etc.
 INT16 gsCreatureInsertionCode = 0;
 INT32 gsCreatureInsertionGridNo = 0;
-UINT8 gubNumCreaturesAttackingTown = 0;
-UINT8 gubYoungMalesAttackingTown = 0;
-UINT8 gubYoungFemalesAttackingTown = 0;
-UINT8 gubAdultMalesAttackingTown = 0;
-UINT8 gubAdultFemalesAttackingTown = 0;
+UINT16 gubNumCreaturesAttackingTown = 0;
+UINT16 gubYoungMalesAttackingTown = 0;
+UINT16 gubYoungFemalesAttackingTown = 0;
+UINT16 gubAdultMalesAttackingTown = 0;
+UINT16 gubAdultFemalesAttackingTown = 0;
 UINT8 gubCreatureBattleCode = CREATURE_BATTLE_CODE_NONE;
 UINT8 gubSectorIDOfCreatureAttack = 0;
 
@@ -1634,8 +1634,8 @@ BOOLEAN MineClearOfMonsters( UINT8 ubMineIndex )
 	return TRUE;
 }
 
-void DetermineCreatureTownComposition( UINT8 ubNumCreatures, UINT8 *pubNumYoungMales, UINT8 *pubNumYoungFemales,
-									  UINT8 *pubNumAdultMales, UINT8 *pubNumAdultFemales )
+void DetermineCreatureTownComposition( UINT16 ubNumCreatures, UINT16 *pubNumYoungMales, UINT16 *pubNumYoungFemales,
+									  UINT16 *pubNumAdultMales, UINT16 *pubNumAdultFemales )
 {
 	INT32 i, iRandom;
 	UINT8 ubYoungMalePercentage = 10;
@@ -1667,8 +1667,8 @@ void DetermineCreatureTownComposition( UINT8 ubNumCreatures, UINT8 *pubNumYoungM
 	}
 }
 
-void DetermineCreatureTownCompositionBasedOnTacticalInformation( UINT8 *pubNumCreatures, UINT8 *pubNumYoungMales, UINT8 *pubNumYoungFemales,
-																UINT8 *pubNumAdultMales, UINT8 *pubNumAdultFemales )
+void DetermineCreatureTownCompositionBasedOnTacticalInformation( UINT16 *pubNumCreatures, UINT16 *pubNumYoungMales, UINT16 *pubNumYoungFemales,
+																UINT16 *pubNumAdultMales, UINT16 *pubNumAdultFemales )
 {
 	SOLDIERTYPE *pSoldier;
 
@@ -1704,7 +1704,7 @@ void DetermineCreatureTownCompositionBasedOnTacticalInformation( UINT8 *pubNumCr
 	}
 }
 
-void DetermineOtherCreatureTownCompositionBasedOnTacticalInformation( UINT8* pubNumCreatures, UINT8* pubNumBloodcats, UINT8* pubNumZombies, UINT8* pubNumBandits )
+void DetermineOtherCreatureTownCompositionBasedOnTacticalInformation( UINT16* pubNumCreatures, UINT16* pubNumBloodcats, UINT16* pubNumZombies, UINT16* pubNumBandits )
 {
 	SOLDIERTYPE *pSoldier;
 
@@ -1750,13 +1750,13 @@ BOOLEAN PrepareCreaturesForBattle()
 	UINT8 ubAdultMalePercentage;
 	UINT8 ubAdultFemalePercentage;
 	UINT8 ubCreatureHabitat;
-	UINT8 ubNumLarvae = 0;
-	UINT8 ubNumInfants = 0;
-	UINT8 ubNumYoungMales = 0;
-	UINT8 ubNumYoungFemales = 0;
-	UINT8 ubNumAdultMales = 0;
-	UINT8 ubNumAdultFemales = 0;
-	UINT8 ubNumCreatures;
+	UINT16 ubNumLarvae = 0;
+	UINT16 ubNumInfants = 0;
+	UINT16 ubNumYoungMales = 0;
+	UINT16 ubNumYoungFemales = 0;
+	UINT16 ubNumAdultMales = 0;
+	UINT16 ubNumAdultFemales = 0;
+	UINT16 ubNumCreatures;
 
 	if( !gubCreatureBattleCode )
 	{
